+++ BUILD.gn
@@ -184,11 +184,11 @@ asar("js2asar") {
   }
   if (enable_view_api) {
     sources += [
-      "lib/browser/api/box-layout.js",
-      "lib/browser/api/button.js",
-      "lib/browser/api/label-button.js",
-      "lib/browser/api/layout-manager.js",
-      "lib/browser/api/text-field.js",
+      "lib/browser/api/views/box-layout.js",
+      "lib/browser/api/views/button.js",
+      "lib/browser/api/views/label-button.js",
+      "lib/browser/api/views/layout-manager.js",
+      "lib/browser/api/views/text-field.js",
     ]
   }
   outputs = [
@@ -423,16 +423,16 @@ static_library("electron_lib") {
 
   if (enable_view_api) {
     sources += [
-      "atom/browser/api/atom_api_box_layout.cc",
-      "atom/browser/api/atom_api_box_layout.h",
-      "atom/browser/api/atom_api_button.cc",
-      "atom/browser/api/atom_api_button.h",
-      "atom/browser/api/atom_api_label_button.cc",
-      "atom/browser/api/atom_api_label_button.h",
-      "atom/browser/api/atom_api_layout_manager.cc",
-      "atom/browser/api/atom_api_layout_manager.h",
-      "atom/browser/api/atom_api_text_field.cc",
-      "atom/browser/api/atom_api_text_field.h",
+      "atom/browser/api/views/atom_api_box_layout.cc",
+      "atom/browser/api/views/atom_api_box_layout.h",
+      "atom/browser/api/views/atom_api_button.cc",
+      "atom/browser/api/views/atom_api_button.h",
+      "atom/browser/api/views/atom_api_label_button.cc",
+      "atom/browser/api/views/atom_api_label_button.h",
+      "atom/browser/api/views/atom_api_layout_manager.cc",
+      "atom/browser/api/views/atom_api_layout_manager.h",
+      "atom/browser/api/views/atom_api_text_field.cc",
+      "atom/browser/api/views/atom_api_text_field.h",
     ]
   }
