+++ lib/browser/api/module-list.js
@@ -36,10 +36,10 @@ module.exports = [
 
 if (features.isViewApiEnabled()) {
   module.exports.push(
-    { name: 'BoxLayout', file: 'box-layout' },
-    { name: 'Button', file: 'button' },
-    { name: 'LabelButton', file: 'label-button' },
-    { name: 'LayoutManager', file: 'layout-manager' },
-    { name: 'TextField', file: 'text-field' }
+    { name: 'BoxLayout', file: 'views/box-layout' },
+    { name: 'Button', file: 'views/button' },
+    { name: 'LabelButton', file: 'views/label-button' },
+    { name: 'LayoutManager', file: 'views/layout-manager' },
+    { name: 'TextField', file: 'views/text-field' }
   )
 }
