+++ lib/browser/api/views/box-layout.js
