+++ lib/browser/api/views/label-button.js
