+++ lib/browser/api/views/layout-manager.js
